part-5-observability/solution/requirements.txt
strands-agents[otel]
strands-agents-tools
